MYR-2221: fixed translation doesn't change in login page
